requirements.txt
# Core frameworks
langgraph>=0.0.19
langchain>=0.0.335
langchain-openai>=0.0.2
langchain-community>=0.0.10

# LlamaIndex and components
llama-index==0.12.40
llama-index-core==0.12.40
llama-index-llms-openai==0.4.2
llama-index-embeddings-openai==0.3.1
llama-index-readers-file==0.4.8
llama-index-agent-openai==0.4.9
llama-index-llms-langchain==0.4.4

# Repository management
python-gitlab>=3.15.0
gitpython>=3.1.35

# Utilities
pydantic>=2.4.2
typing-extensions>=4.5.0
typing-inspect>=0.8.0
aiohttp>=3.8.4
requests>=2.28.2
tqdm>=4.65.0
numpy>=1.24.2
nltk>=3.8.1

# Optional dependencies for ML (comment out if not needed)
# pytorch>=2.0.0
# tensorflow>=2.12.0; platform_system != "Darwin" or platform_machine != "arm64"
# tensorflow-macos>=2.12.0; platform_system == "Darwin" and platform_machine == "arm64"
